server/requirements.txt
blinker==1.9.0
charset-normalizer==3.4.3
click==8.1.8
dnspython==2.7.0
ExifRead==3.5.1
Flask==3.1.2
flask-cors==6.0.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.0.2
pandas==2.3.2
pillow==11.3.0
pymongo==4.15.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.2
reportlab==4.4.4
scapy==2.6.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
